\name{bayesLife-internal}
\alias{bdem.parameter.traces}
\alias{e0.DLisDecrement}
\alias{e0.get.dlcurves}
\alias{get.countries.index.bayesLife.mcmc.meta}
\alias{get.data.matrix.bayesLife.mcmc.meta}
\alias{get.data.imputed.bayesLife.prediction}
\alias{get.e0.shift}
\alias{get.friendly.variant.names.bayesLife.prediction}
\alias{get.new.e0.colnames}
\alias{get.nr.countries.bayesLife.mcmc.meta}
\alias{get.nrest.countries.bayesLife.mcmc.meta}
\alias{get.projection.summary.header.bayesLife.prediction}
\alias{get.traj.ascii.header.bayesLife.mcmc.meta}
\alias{get.UN.variant.names.bayesLife.prediction}
\alias{store.bayesLife.convergence}
\alias{store.bayesLife.prediction}

%datasets
\alias{loess.sd}

\title{
Internal Functions and datasets of bayesLife
}

\description{
These functions and datasets are not to be used directly by the user.
}

\keyword{internal}
